New adm command Moved 4 get subcommands to adm Changed a few help texts
