+++ Makefile
@@ -1,4 +1,4 @@
-# Copyright 2016 The Kubernetes Authors.
+    # Copyright 2016 The Kubernetes Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
